backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
pillow==10.1.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1